Review fixes: move the volatile-read guard into the shared SideEffects helper
