+++ src/test/java/org/openrewrite/staticanalysis/SimplifyRedundantLogicalExpressionTest.java
@@ -262,6 +262,25 @@ class SimplifyRedundantLogicalExpressionTest implements RewriteTest {
     }
 
     @Test
+    void doNotChangeVolatileFieldRead() {
+        rewriteRun(
+          //language=java
+          java(
+            """
+              class Test {
+                  volatile boolean flag;
+
+                  @SuppressWarnings("all")
+                  boolean test() {
+                      return flag && flag;
+                  }
+              }
+              """
+          )
+        );
+    }
+
+    @Test
     void simplifyLogicalAndKotlin() {
         rewriteRun(
           //language=kotlin
